refactor(components): split field guard analysis
